DEV-9834; moves time period module into the location of the app it relates to
