--- usaspending_api/common/filters/time_period/__init__.py
@@ -1,3 +1,0 @@
-from .transaction_search import TransactionSearchTimePeriod
-
-__all__ = ["TransactionSearchTimePeriod"]
