+++ usaspending_api/search/filters/time_period/__init__.py
@@ -1,0 +1,1 @@
+from .time_period import AbstractTimePeriod
