+++ usaspending_api/search/filters/time_period/query_type/__init__.py
@@ -1,0 +1,3 @@
+from .transactions import TransactionSearchTimePeriod
+
+__all__ = ["TransactionSearchTimePeriod"]
